python-examples/requirements.txt
# OpenAI Python API Requirements
# pip install -r requirements.txt

openai>=1.12.0
python-dotenv>=1.0.0
requests>=2.31.0
rich>=13.7.0
